[base] Pull request #1763: Feature/LINUXDAR-5425 add an option to watchdog to generate a coredump for stuck plugins
Merge in LINUXEP/everest-base from feature/LINUXDAR-5425-add-an-option-to-watchdog-to-generate-a-coredump-for-stuck-plugins to develop

Squashed commit of the following:

commit 8654711bbf5fe3b12ed70a60d18e6ae3e744f56c
Merge: 52f10c3e0 953d2e579
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Fri Oct 14 14:49:58 2022 +0100

    Merge branch 'develop' of ssh://stash.sophos.net:7999/linuxep/everest-base into feature/LINUXDAR-5425-add-an-option-to-watchdog-to-generate-a-coredump-for-stuck-plugins

commit 52f10c3e0f82198b282538a06b8915c151bb2d8b
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Fri Oct 14 10:03:23 2022 +0100

    [LINUXDAR-5425] setup copy of core dump on aws

commit e9c678f74b684974bf4e7d290173511ba4639ab3
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Tue Oct 11 13:27:29 2022 +0100

    [LINUXDAR-5425] fix unit test 2nd attempt

commit 2f37531f626d174dff8fc7909885416fa6892037
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Tue Oct 11 10:30:54 2022 +0100

    [LINUXDAR-5425] fix unit test

commit 140d3beedd65e4747b744060be8d92142b91fb38
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Mon Oct 10 16:45:04 2022 +0100

    [LINUXDAR-5425] fix unit test

commit 539aad2c8499d259fba0549e3e3feabefafa2a4f
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Fri Oct 7 11:19:26 2022 +0100

    [LINUXDAR-5425]fix test teardown

commit 55d2b82553fc6fa40a9e931d58ec0b0294db8006
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Fri Oct 7 11:02:29 2022 +0100

    [LINUXDAR-5425] add unit test

commit 190281124214a7697074c808f4b15bed9edb3a4f
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Thu Oct 6 17:14:57 2022 +0100

    [LINUXDAR-5425] when env variable is set make watchdog kill plugins with sigabrt instead of sigkill

commit 2cb18f53e8c401f7107142dfe0f719f20723f47e
Author: Menaga Karunanandan <menaga.karunanandan@sophos.com>
Date:   Thu Oct 6 14:29:27 2022 +0100

    [LINUXDAR-5425] add method to send sigabrt when killing processes

+++ base/modules/Common/ProcessImpl/ProcessImpl.cpp
@@ -71,6 +71,7 @@ namespace Common
             bool hasFinished() override { return true; }
 
             void sendTerminateSignal() override { return; }
+            void sendAbortSignal() override { return; }
 
             void kill() override { return; }
         };
@@ -98,6 +99,7 @@ namespace Common
             bool hasFinished() override { return true; }
 
             void sendTerminateSignal() override { return; }
+            void sendAbortSignal() override { return; }
 
             void kill() override { return; }
         };
@@ -204,7 +206,15 @@ namespace Common
                 processOnBoost->sendTerminateSignal();
                 if (wait(Process::milli(numOfDecSeconds), 100) == Process::ProcessStatus::TIMEOUT)
                 {
-                    processOnBoost->kill();
+                    if (m_coreDumpEnabled)
+                    {
+                        processOnBoost->sendAbortSignal();
+                    }
+                    else
+                    {
+                        processOnBoost->kill();
+                    }
+
                     requiredKill = true;
                 }
             }
@@ -264,6 +274,11 @@ namespace Common
         {
             auto processOnBoost = safeAccess();
             processOnBoost->wait();
+        }
+
+        void ProcessImpl::setCoreDumpMode(const bool mode)
+        {
+            m_coreDumpEnabled = mode;
         }
 
         void ProcessImpl::setOutputTrimmedCallback(std::function<void(std::string)> outputTrimmedCallback)
